requirements.txt
allure-pytest==2.8.17
allure-python-commons==2.8.17
amqp==2.5.2
asgiref==3.2.7
atomicwrites==1.4.0
attrs==19.3.0
billiard==3.6.3.0
celery==4.4.2
colorama==0.4.3
cycler==0.10.0
Django==3.0.5
django-grappelli==2.14.1
dnspython==1.16.0
enum34==1.1.10
eventlet==0.25.1
greenlet==0.4.15
iniconfig==1.0.1
kiwisolver==1.1.0
kombu==4.6.8
loguru==0.5.3
lxml==4.5.2
matplotlib==3.1.2
monotonic==1.5
more-itertools==8.4.0
namedlist==1.7
numpy==1.18.1
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.6
pyTelegramBotAPI==3.6.7
pytest==6.0.1
pytest-allure-adaptor==1.7.10
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
redis==3.4.1
requests==2.11.1
requests-toolbelt==0.7.0
selenium==3.141.0
six==1.14.0
sqlparse==0.3.0
toml==0.10.1
urllib3==1.25.8
vine==1.3.0
win32-setctime==1.0.2
